backend/requirements.txt
alembic==1.17.2
aniso8601==10.0.1
attrs==25.4.0
bcrypt==5.0.0
black==25.12.0
blinker==1.9.0
click==8.3.1
ecdsa==0.19.1
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
flask-marshmallow==1.3.0
flask-restx==1.3.2
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
importlib_resources==6.5.2
iniconfig==2.3.0
isort==7.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
loguru==0.7.3
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.1.1
mdurl==0.1.2
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.5.1
pluggy==1.6.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.2
pytest==9.0.2
python-dotenv==1.2.1
python-jose==3.5.0
pytokens==0.3.0
PyYAML==6.0.3
referencing==0.37.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.45
typing_extensions==4.15.0
Werkzeug==3.1.4
